Moving the openapi implementation into one class
This fixes the 405 error being a 404 error when there are multiple classes implementing
an openapi yaml. This doesn't seem too terrible since it is still possible to separate
the filter stuff and the non-filter stuff into separate classes.
