+++ .gitignore
@@ -1,2 +1,3 @@
 target
-
+.idea
+.vscode
