--- src/main/resources/openapi/sign-token-1.0.yaml
@@ -1,55 +1,0 @@
-openapi: 3.0.0
-info:
-  title: Tenant API
-  version: v2
-paths:
-  /token:
-    parameters:
-      # TODO Should these be here ore in the yaml for AuthorizeApi?
-      - $ref: headers/okapi-tenant.yaml
-      - $ref: headers/okapi-token.yaml
-    post:
-      description: Sign token
-      operationId: token
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: "#/components/schemas/signTokenPayload"
-        required: true
-      responses:
-        "201":
-          description: Created and signed token successfully
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/token"
-        "400":
-          $ref: "#/components/responses/trait_400"
-        "500":
-          $ref: "#/components/responses/trait_500"
-components:
-  responses:
-    trait_400:
-      description: Bad request
-      content:
-        text/plain:
-          schema:
-            type: string
-            example: Invalid JSON in request
-        application/json:
-          schema:
-            type: object
-            example: { "error": "Invalid JSON in request" }
-    trait_500:
-      description: Internal error
-      content:
-        text/plain:
-          schema:
-            type: string
-            example: Internal server error, contact administrator
-  schemas:
-    signTokenPayload:
-      $ref: schemas/signTokenPayload.json
-    token:
-      $ref: schemas/token.json
